Aura Berserk Fix
I am still learning lua and got some help to understand how to fix this but I finally got it to work in game. One part of it was missing a = which was the easy part. Although what was causing the lua to break was the lack of values for powerMod, protectionMod, and critMod on Init, which has been added, otherwise was coming up with a nil.

I made sure to examine the file under update to make sure the values I put under init were correct.