Support for Inkling mod planet type
Added "flooded" and "floodedoceanfloor" biomes to the infinite oxygen buffs and effects. 
Flooded planets are essentially an alternate Oceanic planet found in the Inkling species mod, meant to be where they natively live, so it makes logical sense.

Might remove this change later on if I ever merge Flooded planet changes back with official Oceanic planets, but it's not planned currently.